move button to send dialog
